Combine tests and streamline implementation
